Break out the `version` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`version` command.
